requirements.txt
pandas
pillow
matplotlib
tqdm
pycocotools
opencv-python
requests
torch
torchvision
scikit-image
kagglehub